refactor: move single post to app dir
