+++ src/app/layout.tsx
@@ -6,6 +6,7 @@ import Analytics from '@/components/analytics'
 import Footer from '@/components/footer'
 import Navbar from '@/components/navbar'
 import Providers from '@/components/providers'
+import ScrollObserver from '@/components/scroll-observer'
 
 const fontSans = Poppins({
   subsets: ['latin'],
@@ -22,14 +23,16 @@ export default function RootLayout({
     <html lang="en" suppressHydrationWarning className={fontSans.variable}>
       <body className="antialiased text-black bg-slate-100 dark:bg-gray-900 dark:text-slate-100">
         <Providers>
-          <Navbar />
-          <main
-            className="flex flex-col mx-auto max-w-5xl justify-center py-10 px-4"
-            role="main"
-          >
-            <div className="min-h-screen">{children}</div>
-          </main>
-          <Footer />
+          <ScrollObserver>
+            <Navbar />
+            <main
+              className="flex flex-col mx-auto max-w-5xl justify-center py-10 px-4"
+              role="main"
+            >
+              <div className="min-h-screen">{children}</div>
+            </main>
+            <Footer />
+          </ScrollObserver>
           <Analytics />
         </Providers>
       </body>
