+++ src/components/mdx/image.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import classnames from 'classnames'
 import type { ImageProps as NextImageProps } from 'next/image'
 import NextImage from 'next/image'
