+++ src/components/mdx/views-counter.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import { useEffect } from 'react'
 
 import { useRequest } from '@/hooks'
