+++ src/components/scroll-progress-bar.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import { useContext, useEffect, useState } from 'react'
 
 import { ScrollContext } from './scroll-observer'
